Conflicts:
	README.md
	flow/core/src/main/java/br/net/mirante/singular/flow/core/ProcessDataServiceImpl.java
	flow/core/src/main/java/br/net/mirante/singular/flow/core/ProcessDefinition.java
	flow/core/src/main/java/br/net/mirante/singular/flow/core/service/IUIAdminService.java
	flow/ui-admin/src/main/java/br/net/mirante/singular/dao/InstanceDAO.java
	flow/ui-admin/src/main/java/br/net/mirante/singular/service/PesquisaService.java
	flow/ui-admin/src/main/java/br/net/mirante/singular/service/UIAdminFacade.java
	flow/ui-admin/src/main/java/br/net/mirante/singular/view/error/Error403Page.utf8.properties
	flow/ui-admin/src/main/java/br/net/mirante/singular/view/error/Error403Page_pt_BR.utf8.properties
	flow/ui-admin/src/main/java/br/net/mirante/singular/view/page/dashboard/DashboardContent.java
	flow/ui-admin/src/main/java/br/net/mirante/singular/view/page/dashboard/FeedPanel.java
	flow/ui-admin/src/main/java/br/net/mirante/singular/wicket/UIAdminSession.java

+++ flow/core/src/main/java/br/net/mirante/singular/flow/core/service/IProcessDataService.java
@@ -15,7 +15,7 @@ import br.net.mirante.singular.flow.core.entity.IEntityTaskDefinition;
  */
 public interface IProcessDataService<I extends ProcessInstance> {
 
-    I retrieveInstance(Integer entityCod);
+    I retrieveInstance(Long entityCod);
 
     List<I> retrieveActiveInstancesCreatedBy(MUser user);
